Fix return value type of vde_send().
The libvdeplug_dyn.h header file incorrectly declares the return type of
vde_send() to size_t, while in reality it is ssize_t.

+++ src/vde_device.c
@@ -114,7 +114,7 @@ static bool read_packet(vpn_packet_t *packet) {
 }
 
 static bool write_packet(vpn_packet_t *packet) {
-	if(plug.vde_send(conn, packet->data, packet->len, 0) < 0) {
+	if((ssize_t)plug.vde_send(conn, packet->data, packet->len, 0) < 0) {
 		if(errno != EINTR && errno != EAGAIN) {
 			logger(LOG_ERR, "Can't write to %s %s: %s", device_info, device, strerror(errno));
 			running = false;
